+++ src/sys/util.cpp
@@ -67,8 +67,9 @@ std::vector<std::string> Settings::parseArgs(int argc, char *argv[]) {
       std::stringstream s(arg);
       int64_t val;
       s >> val;
-      if (option != Option::MAX_CYCLES && option != Option::MAX_EVAL_SECS &&
-          option != Option::B_FILE_OFFSET && val < 1) {
+      if (option != Option::MAX_CYCLES && option != Option::MAX_MEMORY &&
+          option != Option::MAX_EVAL_SECS && option != Option::B_FILE_OFFSET &&
+          val < 1) {
         Log::get().error("Invalid value for option: " + std::to_string(val),
                          true);
       }