+++ CHANGELOG.md
@@ -5,6 +5,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 ### Bugfixes
 
 * Fix optimizer `prg`-handling
+* Fix max memory option`-m`
 
 ### Enhancements